test_uce/pyproject.toml
[project]
name = "test-uce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.4",
    "groq>=0.16.0",
    "loguru>=0.7.3",
    "openai>=1.60.2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]

[tool.uv.workspace]
members = ["chat_bot"]
